requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.11.1
altair==5.2.0
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
Babel==2.12.1
bcrypt==4.0.1
billiard==4.1.0
bleach==6.1.0
blinker==1.6.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
distro==1.9.0
dnspython==2.3.0
dominate==2.8.0
elastic-transport==8.4.0
elasticsearch==8.9.0
email-validator==2.0.0.post2
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
gradio==3.5
gradio_client==0.8.1
greenlet==2.0.2
h11==0.12.0
httpcore==0.15.0
httpx==0.25.1
huggingface-hub==0.20.3
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.3.2
langdetect==1.0.9
linkify-it-py==2.0.2
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mathparse==0.1.2
matplotlib==3.8.2
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
orjson==3.9.12
packaging==23.1
pandas==2.2.0
paramiko==3.4.0
passlib==1.7.4
pdfkit==1.0.0
Pillow==9.5.0
Pint==0.23
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pymongo==3.13.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
ruff==0.1.14
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
speaklater==1.3
SQLAlchemy==2.0.25
starlette==0.35.1
stripe==7.13.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.27.0
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.8
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.4
